requirements.txt
flask==2.3.2
openai==1.35.7
python-dotenv==1.0.1